--- api-kmp/src/commonMain/kotlin/IApiStrategy.kt
@@ -1,11 +1,0 @@
-package aigt.api.kmp.v1
-
-import kotlinx.serialization.KSerializer
-import kotlin.reflect.KClass
-
-interface IApiStrategy<K: Any> {
-    val discriminator: String
-    val clazz: KClass<out K>
-    val serializer: KSerializer<out K>
-    fun <T: K> fillDiscriminator(req: T): T
-}
