--- api-kmp/src/commonMain/kotlin/requests/HistoryRequestStrategy.kt
@@ -1,17 +1,0 @@
-package aigt.api.kmp.v1.requests
-
-import kotlinx.serialization.KSerializer
-import aigt.api.kmp.v1.models.AccountHistoryRequest
-import aigt.api.kmp.v1.models.IRequest
-import kotlin.reflect.KClass
-
-object HistoryRequestStrategy: IRequestStrategy {
-    override val discriminator: String = "history"
-    override val clazz: KClass<out IRequest> = AccountHistoryRequest::class
-    override val serializer: KSerializer<out IRequest> = AccountHistoryRequest.serializer()
-    override fun <T : IRequest> fillDiscriminator(req: T): T {
-        require(req is AccountHistoryRequest)
-        @Suppress("UNCHECKED_CAST")
-        return req.copy(requestType = discriminator) as T
-    }
-}
