+++ api-v1-jackson/src/main/kotlin/ApiV1Mapper.kt
@@ -1,0 +1,28 @@
+package aigt.api.v1.jackson
+
+import com.fasterxml.jackson.databind.DeserializationFeature
+import com.fasterxml.jackson.databind.MapperFeature
+import com.fasterxml.jackson.databind.json.JsonMapper
+import aigt.api.v1.jackson.models.IRequest
+import aigt.api.v1.jackson.models.IResponse
+
+val apiV1Mapper: JsonMapper = JsonMapper.builder().run {
+    configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
+    enable(MapperFeature.USE_BASE_TYPE_AS_DEFAULT_IMPL)
+//    setSerializationInclusion(JsonInclude.Include.NON_NULL)
+    build()
+}
+
+fun apiV1RequestSerialize(request: IRequest): String =
+    apiV1Mapper.writeValueAsString(request)
+
+@Suppress("UNCHECKED_CAST")
+fun <T : IRequest> apiV1RequestDeserialize(json: String): T =
+    apiV1Mapper.readValue(json, IRequest::class.java) as T
+
+fun apiV1ResponseSerialize(response: IResponse): String =
+    apiV1Mapper.writeValueAsString(response)
+
+@Suppress("UNCHECKED_CAST")
+fun <T : IResponse> apiV1ResponseDeserialize(json: String): T =
+    apiV1Mapper.readValue(json, IResponse::class.java) as T
