+++ api-v1-kmp/build.gradle.kts
@@ -1,0 +1,82 @@
+plugins {
+    kotlin("multiplatform")
+    id("org.openapi.generator")
+    kotlin("plugin.serialization")
+}
+
+kotlin {
+    jvm { }
+    linuxX64 { }
+//    macosX64 { }
+
+    sourceSets {
+        val serializationVersion: String by project
+
+        @Suppress("UNUSED_VARIABLE")
+        val commonMain by getting {
+
+            kotlin.srcDirs("$buildDir/generate-resources/main/src/commonMain/kotlin")
+
+            dependencies {
+                implementation(kotlin("stdlib-common"))
+
+                implementation("org.jetbrains.kotlinx:kotlinx-serialization-core:$serializationVersion")
+                implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:$serializationVersion")
+            }
+        }
+        @Suppress("UNUSED_VARIABLE")
+        val commonTest by getting {
+            dependencies {
+                implementation(kotlin("test-common"))
+                implementation(kotlin("test-annotations-common"))
+            }
+        }
+        @Suppress("UNUSED_VARIABLE")
+        val jvmTest by getting {
+            dependencies {
+                implementation(kotlin("test-junit"))
+            }
+        }
+    }
+}
+
+/**
+ * Настраиваем генерацию здесь
+ */
+openApiGenerate {
+    val openapiGroup = "${rootProject.group}.api.v1.kmp"
+    generatorName.set("kotlin") // Это и есть активный генератор
+    packageName.set(openapiGroup)
+    apiPackage.set("$openapiGroup.api")
+    modelPackage.set("$openapiGroup.models")
+    invokerPackage.set("$openapiGroup.invoker")
+    inputSpec.set("$rootDir/specs/specs-account-v1.yaml")
+    library.set("multiplatform")
+
+    /**
+     * Здесь указываем, что нам нужны только модели, все остальное не нужно
+     */
+    globalProperties.apply {
+        put("models", "")
+        put("modelDocs", "false")
+    }
+
+    /**
+     * Настройка дополнительных параметров из документации по генератору
+     * https://github.com/OpenAPITools/openapi-generator/blob/master/docs/generators/kotlin.md
+     */
+    configOptions.set(
+        mapOf(
+            "dateLibrary" to "string",
+            "enumPropertyNaming" to "UPPERCASE",
+            "collectionType" to "list",
+        )
+    )
+}
+
+afterEvaluate {
+    val openApiGenerate = tasks.getByName("openApiGenerate")
+    tasks.filter { it.name.startsWith("compile") }.forEach {
+        it.dependsOn(openApiGenerate)
+    }
+}
