+++ api-v1-kmp/src/commonMain/kotlin/IApiStrategy.kt
@@ -1,0 +1,11 @@
+package aigt.api.v1.kmp
+
+import kotlinx.serialization.KSerializer
+import kotlin.reflect.KClass
+
+interface IApiStrategy<K: Any> {
+    val discriminator: String
+    val clazz: KClass<out K>
+    val serializer: KSerializer<out K>
+    fun <T: K> fillDiscriminator(req: T): T
+}
