+++ api-v1-kmp/src/commonMain/kotlin/requests/ReadRequestStrategy.kt
@@ -1,0 +1,17 @@
+package aigt.api.v1.kmp.requests
+
+import kotlinx.serialization.KSerializer
+import aigt.api.v1.kmp.models.AccountReadRequest
+import aigt.api.v1.kmp.models.IRequest
+import kotlin.reflect.KClass
+
+object ReadRequestStrategy: IRequestStrategy {
+    override val discriminator: String = "read"
+    override val clazz: KClass<out IRequest> = AccountReadRequest::class
+    override val serializer: KSerializer<out IRequest> = AccountReadRequest.serializer()
+    override fun <T : IRequest> fillDiscriminator(req: T): T {
+        require(req is AccountReadRequest)
+        @Suppress("UNCHECKED_CAST")
+        return req.copy(requestType = discriminator) as T
+    }
+}
