+++ api-v1-kmp/src/commonMain/kotlin/requests/SearchRequestStrategy.kt
@@ -1,0 +1,17 @@
+package aigt.api.v1.kmp.requests
+
+import kotlinx.serialization.KSerializer
+import aigt.api.v1.kmp.models.AccountSearchRequest
+import aigt.api.v1.kmp.models.IRequest
+import kotlin.reflect.KClass
+
+object SearchRequestStrategy: IRequestStrategy {
+    override val discriminator: String = "search"
+    override val clazz: KClass<out IRequest> = AccountSearchRequest::class
+    override val serializer: KSerializer<out IRequest> = AccountSearchRequest.serializer()
+    override fun <T : IRequest> fillDiscriminator(req: T): T {
+        require(req is AccountSearchRequest)
+        @Suppress("UNCHECKED_CAST")
+        return req.copy(requestType = discriminator) as T
+    }
+}
